backend_django/requirements.txt
Django==5.1.1
dj-database-url==2.2.0
environs==11.0.0
psycopg2-binary==2.9.9
djangorestframework==3.15.2
django-phonenumber-field==8.0.0
phonenumbers==8.13.46
djoser==2.2.3
transliterate==1.10.2
pillow==10.4.0
drf-spectacular==0.27.2